Move computer vision to its own class
